Share code for modifications exporting filters
